requirements-enterprise.txt
# Enterprise SaaS Platform Dependencies

# FastAPI and ASGI
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy[asyncio]==2.0.23
asyncpg==0.29.0
alembic==1.13.1

# MongoDB
motor==3.3.2
pymongo==4.6.0

# Redis
redis[hiredis]==5.0.1

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
cryptography==41.0.8

# Data validation
pydantic[email]==2.5.2
pydantic-settings==2.1.0

# HTTP client
httpx==0.25.2
aiohttp==3.9.1

# Utilities
python-dotenv==1.0.0
structlog==23.2.0
rich==13.7.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Development
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1

# Monitoring
prometheus-client==0.19.0
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0

# Background tasks
celery[redis]==5.3.4

# File handling
aiofiles==23.2.1

# Data processing (keeping some from original)
pandas==2.3.1
numpy==1.26.4
scikit-learn==1.7.0
joblib==1.5.1
openpyxl==3.1.2  # Excel file support
xlrd==2.0.1      # Excel file support (legacy)
chardet==5.2.0   # Character encoding detection
lxml==4.9.3      # XML processing
pyarrow==14.0.1  # Parquet file support

# Keep existing ML libraries for compatibility
xgboost==3.0.2
statsmodels==0.14.5
prophet==1.1.7